add ArmAgentController
